SCM/Backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via
    #   django
    #   django-cors-headers
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via pytest
django==4.2.7
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-extensions
    #   djangorestframework
django-cors-headers==4.3.1
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
geojson==3.1.0
    # via -r requirements.in
idna==3.6
    # via requests
iniconfig==2.0.0
    # via pytest
joblib==1.3.2
    # via scikit-learn
numpy==1.26.4
    # via
    #   -r requirements.in
    #   pandas
    #   scikit-learn
    #   scipy
openrouteservice==2.3.3
    # via -r requirements.in
packaging==23.2
    # via pytest
pandas==2.2.0
    # via -r requirements.in
pluggy==1.4.0
    # via pytest
psycopg2==2.9.9
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pytest==7.4.3
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   pandas
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2024.1
    # via
    #   djangorestframework
    #   pandas
requests==2.31.0
    # via
    #   -r requirements.in
    #   openrouteservice
scikit-learn==1.3.0
    # via -r requirements.in
scipy==1.12.0
    # via scikit-learn
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
threadpoolctl==3.4.0
    # via scikit-learn
tzdata==2024.1
    # via
    #   django
    #   pandas
urllib3==2.2.1
    # via requests
